requirements.txt
Flask==3.0.3
Pillow==11.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
gunicorn==21.2.0
